+++ app/contact/Address.tsx
@@ -1,0 +1,8 @@
+
+export default function Address() {
+    return (<address>
+        Amaris Homes Subdivision Phase 2, Burol Main, Dasmarinas City,{" "}
+        <br className="d-none d-md-block" />
+        Cavite, Philippines 4114
+    </address>);
+}
