+++ app/contact/ContactMap.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+export default function ContactMap() {
+    return (<iframe
+        src="https://www.google.com/maps/embed?pb=!1m14!1m12!1m3!1d1930.8155887771835!2d120.944905!3d14.335279!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!5e0!3m2!1sen!2sph!4v1718127890000!5m2!1sen!2sph"
+        style={{ width: "100%", height: "100%", border: 0 }}
+        allowFullScreen={true}
+    />);
+}
