+++ app/contact/PhoneNumber.tsx
@@ -1,0 +1,7 @@
+
+export default function PhoneNumber() {
+    return (<p>
+        +63 966 684 1752 <br className="d-none d-md-block" />
+        +63 917 872 1024
+    </p>);
+}
