+++ app/page.tsx
@@ -2,8 +2,8 @@ import CompanyProfile from "./home/CompanyProfile";
 import Hero from "./home/Hero";
 import Services from "./home/Services";
 import Team from "./home/Team";
-import Contact from "./home/Contact";
-import ContactForm from "./home/ContactForm";
+import Contact from "./contact/Contact";
+import ContactDialog from "./contact/ContactDialog";
 
 export default function Page() {
   return (<>
@@ -17,7 +17,7 @@ export default function Page() {
         <div className="modal-content">
           <div className="modal-body">
             <button type="button" className="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            <ContactForm />
+            <ContactDialog />
           </div>
         </div>
       </div>
